implement our own CurrencyToVoteHandler to better match our currency configuration
